Fix solved problems not updating on team join
Cache was not invalidated for list of solved problems and score for both
joining player and team after `join_team` event, now forces refresh. The
scoreboard updates to reflect the changes per the stats caching daemon,
every 60s.

Original issue reproduction:
1. Create account and `team1` as `leader1`.
2. Create account as `player2` without joining team.
3. Solve `problem-A` as `player2`
4. Join `team1` as `player2`.

At this point, all members of `team1` do not see `problem-A` as solved. This
would not be resolved until a new correct flag is able to be submitted (past
cached problem-locked and already-solved limitations). In addition, the
scoreboard would continue to show team `player2` with its prejoin score.