Clarify error messages when join/create teams
Specifically that a team cannot be created that conflicts with
an existing username. Likewise, an attempt to join a team that does
conflict with an existing username (which is in reality, a user's
starting self-team) receives the same error. The user check is
performed first when validating a team creation so that the
later team name check and "Team already exists" message does
not create confusion.