Disable PasswordAuthentication by default
This change is potentially not admin friendly because if an admin let
this and the next (restart sshd) provision steps run and fails to insert
an authorized key before logging out, then there is no way to get back
in. However, both vagrant and AWS by default will insert an authorized
key for the admin user and therefore this change is reasonable.

This also makes a certain picoCTF problem more meaningful.