First pass at installing problems from source
Allows passing a directory to 'shell_manager deploy' rather than
a service name (similar to the dry-run flag behavior.)
Behind the scenes, still creates and installs a .deb package to
handle problem dependencies.
This is a first pass at this, some refactoring of shell_manager/package
will be needed in order to clean it up.