Add darglint and bandit linter to CI config
Adds the darglint checks to the flake8 CI job, to check that docstrings
matches Google style guide. It integrates to flake8, you only have to
install the packge and you get the additional checks in directly in
flake8.

Adds the bandit security linter as another CI job. At the moment, the
default configuration is used, but we should use a custom config. The
job is currently allowed to fail.