Add initial CI config for picoCTF-shell tests and some unit tests
Adds a Tox environment to run picoCTF-shell tests with code coverage.
The old shell tests are set to "skip", because they are not working
at the moment. The new Tox environment is added to the Travis-CI
configuration to run the tests on Travis and upload the coverage to
codecov.

Adds a "requirements.txt" to "picoCTF-shell" with all dependencies of
the shell_manager and the hacksport package. It's used to install the
dependencies before the test run on Travis-CI.