Fix Nginx unreliable test of referrer for reverse proxy
This broke the catch of absolute links to root in web problems, for
instance with a form `action="/login"` or a flask `redirect('/')`

Yet unable to determine why it stopped working, but initial tests
were not done behind an AWS ELB, might be related. A more lax
regex expression works in that environment, and shouldn't stop
working elsewhere.

- Fix spacing on nginx template
- Use reverse proxy url scheme in example web exploitation problem