Check 'blacklist' when creating users
Issue:
The `picoCTF` web interface has a configuration box for blacklisted
usernames and a function that checks a username against that list.
However, the check function was not called during the user creation
process.  This lack of checking allows registrations with `root`
and `hacksports` to occur.

Note:
This bug is not a catastophic vulnerability because the current
configuration of PAM forces local authentication for local users.
It currently results in these users being unable to log into the
shell server because their password is checked against the shadow
file and not the user database.

Changes:
- Renames the blacklist check function to make its semantics clear.
- Adds the check to `add_user` and returns the "There is already a
user with this username" message when in the blacklist.  We chose
to reuse this message rather than an explicit blacklist message to
avoid potential leakage of system-level usernames to unregistered
users (registered usernames could log onto shell server and look at
`/etc/passwd`).