Close bypass method for email whitelists
Issue:
The regex used for comparing registration emails to a provided
whitelist is vulnerable to a stuffing attack when paired with the
verification email process.  Specifically, an attacker can enter
an email in the form `bad@bad.com;good@good.com` which will pass
a domain check for `good.com` but send the verification email to
`bad@bad.com`.

Note:
The modified regex now forces the entire portion after the first
`@` to match the admin provided regex.  This is not strictly RFC
compliant (`@` symbol is allowed in quoted local part), but is
effective and should not cause issues for non-ASCII character sets
either.

Changes:
- Modify regex for local part to always terminate at the `@` symbol.