Prevent logging of plaintext passwords
**Potentially introduces breaking change for analysis tools that
run against the `statistics` collection (none seen in repo).**

Issue:
The `statistics` log inadvertently captures plaintext passwords at
several points in the logging process (most noticeably at
registration, login, password reset, and password change).  This
does not follow best practices and defeats the purpose of storing
hashed passwords in the `users` collection.

Note:
This creates breaking change in the schema of the `statistics`
collection by removing the `kwargs` field and changing the `args`
field to a dictionary rather than the previous list.  However, the
end result is more useful for debugging (all parameters identified
by name) and removes all occurrences of plaintext passwords stored
in the Mongo database.

Changes:
- Modified `log_action` decorator to now take in list of parameters
to ignore when going to log.
- Modified logger (`log_action` and `FunctionLoggingHandler`) to
unify `args` and `kwargs` using the `inspect.getcallargs` function
in Python.  This makes the censoring process significantly easier
and has the added benefit of clearly labeling the parameter list in
the logs.
- Updated all uses of `@log_action` to `@log_action()` or to use
the new `dont_log` functionality (all sites where a plaintext
password is used).