Make web API check for docker_challenge more robust
In the existing tests, the hard coded `sample_shellserver_publish_output`
does not contain the docker_challenge field.  This modifications is a
better test to ensure that previously deployed shell manager challenges
are still functional after the inclusion of the DockerChallenge
functionality.

TODO: update sample_shellserver_publish_output and include tests for
DockerChallenge