Remove legacy vagrant configurations
Out-of-date, unsupported and not necessary to provide. Only adds
additional complexity and confusion for someone getting started.

Those familiar with Vagrant, who might use these alternative
configurations, are more likely to have unique needs and constraints
that would result in them simply crafting an alternative setup that
suits their needs.

This further cements the 2-machine setup as the default and only
supported configuration.