Pass WIP/SIP environment variables into vagrant vms
This is necessary to ensure a standard experience when using WIP/SIP
environment variables on the host machine as well as in the guest
machine (e.g.  with the ansible_local provisioner).

The following line from infra_local/inventory.yml demonstrates where the
issue was:

```
ansible_host                : "{{ lookup('env','SIP') or '192.168.2.3' }}"
```

The lookup 'env' plugin:

> query the environment variables available on the controller when you
> invoked Ansible.

When using the `ansible_local` provisioner the controller is the guest
virtual machine, not the host where you are setting the ENV var on the
cli.