Kill instance processes by username in undeploy
Additionally, do not log an error when there are no processes to kill
(previously, the fuser call would return 1 if no processes matched)

Also prevents undeploy from killing the parent xinetd process.
In practice, this was usually not an issue since it gets restarted at the
end of undeploy_problems().